requirements-python.txt
# Python Security Scanner - All Dependencies
# Install with: pip install -r requirements-python.txt

# FastAPI & Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.40
aiosqlite==0.21.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.22

# HTTP & Networking
httpx==0.28.1
requests==2.31.0
aiohttp==3.9.1
dnspython==2.4.2

# Data Processing & Parsing
beautifulsoup4==4.13.3
lxml==4.9.3
jinja2==3.1.6
jsonpath-ng==1.6.1
xmltodict==0.13.0

# Validation & Models
pydantic[email]==2.11.3
pydantic-settings==2.9.1
validators==0.22.0

# YAML & Config
pyyaml==6.0.1

# Performance & Caching
cachetools==5.3.2
tenacity==8.2.3

# Async & Concurrency
asyncio-contextmanager==1.0.0

# CLI & Logging
click==8.1.7
rich==13.7.0
colorama==0.4.6
loguru==0.7.2

# Utilities
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
